🎨 네이밍 변경
